+++ exonum/src/events/noise/wrappers/sodium_wrapper/wrapper.rs
@@ -27,8 +27,8 @@ use std::{
 use super::{handshake::HandshakeParams, resolver::SodiumResolver};
 use events::noise::{error::NoiseError, HEADER_LENGTH, MAX_MESSAGE_LENGTH, TAG_LENGTH};
 
-pub const HANDSHAKE_HEADER_LENGTH: usize = 1;
-pub const MAX_HANDSHAKE_MESSAGE_LENGTH: usize = 255;
+pub const HANDSHAKE_HEADER_LENGTH: usize = 4;
+pub const MAX_HANDSHAKE_MESSAGE_LENGTH: usize = 1024;
 pub const MIN_HANDSHAKE_MESSAGE_LENGTH: usize = 32;
 
 // We choose XK pattern since it provides mutual authentication,
@@ -44,9 +44,9 @@ pub struct NoiseWrapper {
 
 impl NoiseWrapper {
     pub fn initiator(params: &HandshakeParams) -> Self {
-        if let Some(ref remote_key) = params.remote_key {
+        if let Some(ref remote_key) = params.remote_key() {
             let builder: Builder = Self::noise_builder()
-                .local_private_key(params.secret_key.as_ref())
+                .local_private_key(params.secret_key().as_ref())
                 .remote_public_key(remote_key.as_ref());
             let session = builder
                 .build_initiator()
@@ -61,7 +61,7 @@ impl NoiseWrapper {
         let builder: Builder = Self::noise_builder();
 
         let session = builder
-            .local_private_key(params.secret_key.as_ref())
+            .local_private_key(params.secret_key().as_ref())
             .build_responder()
             .expect("Noise session responder failed to initialize");