+++ exonum/src/sandbox/consensus/recovery.rs
@@ -19,7 +19,7 @@ use std::time::Duration;
 
 use crypto::CryptoHash;
 use helpers::{Height, Round, ValidatorId};
-use messages::{PeersRequest, PeersResponse, Precommit, Prevote};
+use messages::{PeersExchange, PeersRequest, Precommit, Prevote};
 use node::{self, ConnectInfo};
 
 use blockchain::ValidatorKeys;
@@ -442,7 +442,15 @@ fn should_restore_peers_after_restart() {
     let (p1, s1, a1) = (sandbox.p(v1), sandbox.s(v1).clone(), sandbox.a(v1));
 
     let peers_request = PeersRequest::new(&p1, &p0, &s1);
-    let peers_response = PeersResponse::new(&p0, &p1, vec![a1], &s0);
+    let peers_response = PeersExchange::new(
+        &p0,
+        &p1,
+        vec![ConnectInfo {
+            public_key: p1,
+            address: a1,
+        }],
+        &s0,
+    );
 
     // check that peers are absent
     sandbox.recv(&peers_request);